SONARPHP-1331 Write/read Project Symbol Tables to/from cache, read from cache
