+++ sonar-php-plugin/src/main/java/org/sonar/plugins/php/PHPSensor.java
@@ -76,13 +76,6 @@ import static org.sonar.plugins.php.warning.DefaultAnalysisWarningsWrapper.NOOP_
 
 public class PHPSensor implements Sensor {
 
-  /**
-   * Describes if an optimized analysis of unchanged files by skipping some rules is enabled.
-   * By default, the property is not set (null), leaving SQ/SC to decide whether to enable this behavior.
-   * Setting it to true or false, forces the behavior from the analyzer independently of the server.
-   */
-  public static final String SONAR_CAN_SKIP_UNCHANGED_FILES_KEY = "sonar.php.skipUnchanged";
-
   private static final Logger LOG = Loggers.get(PHPSensor.class);
   private final FileLinesContextFactory fileLinesContextFactory;
   private final PHPChecks checks;
@@ -167,7 +160,6 @@ public class PHPSensor implements Sensor {
     PHPAnalyzer phpAnalyzer;
 
     private final boolean hasTestFileChecks;
-    private final boolean optimizedAnalysis;
 
     public AnalysisScanner(SensorContext context, ProjectSymbolData projectSymbolData, DurationStatistics statistics) {
       super(context, statistics);
@@ -183,15 +175,9 @@ public class PHPSensor implements Sensor {
         filter(PhpUnitCheck.class::isInstance).
         collect(Collectors.toList());
       hasTestFileChecks = !testFilesChecks.isEmpty();
-      optimizedAnalysis = shouldOptimizeAnalysis();
 
       File workingDir = context.fileSystem().workDir();
       phpAnalyzer = new PHPAnalyzer(allChecks, testFilesChecks, workingDir, projectSymbolData, statistics);
-    }
-
-    private boolean shouldOptimizeAnalysis() {
-      return !(inSonarLint(context)) &&
-        (context.canSkipUnchangedFiles() || context.config().getBoolean(SONAR_CAN_SKIP_UNCHANGED_FILES_KEY).orElse(false));
     }
 
     @Override
